#11: Add Home, Tracks, Albums, and Artists page, show corresponding content on their own page
