--- src/renderer/src/audio/AudioList.tsx
@@ -1,19 +1,0 @@
-import { FC, ReactElement } from 'react';
-
-const AudioList: FC  = (): ReactElement => {
-    //#region Props and States
-    //#endregion
-
-    //#region Methods
-    //#endregion
-
-    //#region useEffects
-    //#endregion
-
-    //#region Render
-    return <>
-    </>;
-    //#endregion
-};
-
-export default AudioList;
