+++ src/renderer/src/home/HomePage.tsx
@@ -1,0 +1,15 @@
+import { Typography } from "@mui/material";
+import { FC, ReactElement } from 'react';
+
+const HomePage: FC = (): ReactElement => {
+
+    //#region Render
+    return <>
+        <Typography variant="h3">
+            Home Page
+        </Typography>
+    </>;
+    //#endregion
+};
+
+export default HomePage;
