+++ src/renderer/src/navigation/Types.ts
@@ -1,0 +1,7 @@
+export enum Page {
+    Home = "home",
+    Search = "search",
+    Tracks = "tracks",
+    Albums = "albums",
+    Artists = "artists",
+}
